Refactored TapChangers implementations in order to avoid code duplication
